Refactor appsec wordpress test suite
