+++ tests/Integrations/WordPress/V5_5/PathParamsTest.php
@@ -2,11 +2,12 @@
 
 namespace DDTrace\Tests\Integrations\WordPress\V5_5;
 
-use DDTrace\Tests\Common\AppsecTestCase;
-use DDTrace\Tests\Frameworks\Util\Request\GetSpec;
-use datadog\appsec\AppsecStatus;
+use DDTrace\Tests\Integrations\WordPress\PathParamsTestSuite;
 
-class PathParamsTest extends AppsecTestCase
+ /**
+ * @group appsec
+ */
+class PathParamsTest extends PathParamsTestSuite
 {
     protected static function getAppIndexScript()
     {
@@ -16,82 +17,5 @@ class PathParamsTest extends AppsecTestCase
     protected function connection()
     {
         return new \PDO('mysql:host=mysql_integration;dbname=test', 'test', 'test');
-    }
-
-    protected function ddSetUp()
-    {
-        parent::ddSetUp();
-        $this->connection()->exec(file_get_contents(__DIR__ . '/../../../Frameworks/WordPress/Version_5_5/wp_2020-10-21.sql'));
-        AppsecStatus::getInstance()->setDefaults();
-    }
-
-    public static function ddSetUpBeforeClass()
-    {
-        parent::ddSetUpBeforeClass();
-        AppsecStatus::getInstance()->init();
-    }
-
-    public static function ddTearDownAfterClass()
-    {
-        AppsecStatus::getInstance()->destroy();
-        parent::ddTearDownAfterClass();
-    }
-
-    public function testPost()
-    {
-        $this->call(
-            GetSpec::create(
-                'Post example',
-                '/hello-world'
-            )
-        );
-
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(1, count($events));
-        $this->assertEquals('hello-world', $events[0]["server.request.path_params"]['name']);
-        $this->assertEquals('push_address', $events[0]['eventName']);
-    }
-
-    public function testCategory()
-    {
-        $this->call(
-            GetSpec::create(
-                'Category',
-                '/category/uncategorized'
-            )
-        );
-
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(1, count($events));
-        $this->assertEquals('uncategorized', $events[0]["server.request.path_params"]['category_name']);
-        $this->assertEquals('push_address', $events[0]['eventName']);
-    }
-
-    public function testAuthor()
-    {
-        $this->call(
-            GetSpec::create(
-                'Author',
-                '/author/test'
-            )
-        );
-
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(1, count($events));
-        $this->assertEquals('test', $events[0]["server.request.path_params"]['author_name']);
-        $this->assertEquals('push_address', $events[0]['eventName']);
-    }
-
-    public function testNonExistingPost()
-    {
-       $this->call(
-            GetSpec::create(
-                'Not existing post',
-                '/non-existing-post'
-            )
-        );
-
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(0, count($events));
     }
 }
